requirements.txt
Flask==2.1.1
Flask-CORS==5.0.1
Werkzeug==2.1.2
PyJWT==2.3.0